requirements.txt
# ── API / servidor ────────────────────────────────────────────────────────────
fastapi==0.115.9
uvicorn==0.34.0
python-multipart==0.0.20

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai==1.82.0
tiktoken==0.9.0

# ── LangChain stack ───────────────────────────────────────────────────────────
langchain==0.3.25
langchain-openai==0.3.16
langchain-chroma==0.2.4
langchain-community==0.3.24
langgraph==0.4.5
langsmith==0.3.45

# ── Vector DB ─────────────────────────────────────────────────────────────────
chromadb==1.0.9

# ── Utilidades ────────────────────────────────────────────────────────────────
python-dotenv==1.1.0
pydantic==2.11.5
requests==2.32.3
python-docx==1.1.2

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.45.1